Added level
contains a basic house layout with a single light switch. I was getting some weird collision errors when the ceiling object was enabled so it is currently set to disabled. There are no lights in the scene as I don't really know what the goal with lighting is at this current time.